refactors to source values from env vars

+++ provisioners/autochrome/install.sh
@@ -1,8 +1,13 @@
 
 sudo apt install -y ruby
 
-cd /tmp
-git clone https://github.com/nccgroup/autochrome.git
-cd autochrome
-rm -rf ~/.config/autochrome/
-sudo -H -u vagrant ruby autochrome.rb
+once="$HOME/.once/autochrome-install"
+mkdir -p "$HOME/.once"
+if [ ! -f "$once" ]; then
+  cd /tmp
+  git clone https://github.com/nccgroup/autochrome.git
+  cd autochrome
+  rm -rf ~/.config/autochrome/
+  sudo -H -u vagrant ruby autochrome.rb -c
+  touch "$once";
+fi
